player屏幕受伤红色闪烁+死亡掉落

+++ Script/HealthBar.cs
@@ -22,7 +22,7 @@ public class HealthBar : MonoBehaviour
     // Update is called once per frame
     void Update()
     {
-        _healthBar.fillAmount = (float)HealthCurrent / HealthMax;
+        _healthBar.fillAmount = (float)HealthCurrent / (float)HealthMax;
         healthText.text = HealthCurrent.ToString() + "/" + HealthMax.ToString();
     }
 }
